recorder refactor
