+++ src/main/recorder.ts
@@ -5,32 +5,13 @@ import * as osn from 'obs-studio-node';
 import { IScene, ISettings } from 'obs-studio-node';
 import WaitQueue from 'wait-queue';
 import { ERecordingFormat } from './obsEnums';
-
-import {
-  writeMetadataFile,
-  runSizeMonitor,
-  deleteVideo,
-  addColor,
-  getSortedVideos,
-  fixPathWhenPackaged,
-  tryUnlinkSync,
-} from './util';
-
-import { Metadata, RecStatus, SaveStatus, VideoQueueItem } from './types';
+import { deleteVideo, addColor, getSortedVideos } from './util';
+import { RecStatus } from './types';
 import { VideoCategory } from '../types/VideoCategory';
 import Activity from '../activitys/Activity';
+import VideoProcessQueue from './VideoProcessQueue';
 
 const { v4: uuidfn } = require('uuid');
-
-const atomicQueue = require('atomic-queue');
-const ffmpeg = require('fluent-ffmpeg');
-const obsRecorder = require('./obsRecorder');
-
-const ffmpegPath = fixPathWhenPackaged(
-  require('@ffmpeg-installer/ffmpeg').path
-);
-
-ffmpeg.setFfmpegPath(ffmpegPath);
 
 type RecorderOptionsType = {
   storageDir: string;
@@ -48,9 +29,6 @@ type RecorderOptionsType = {
   obsRecEncoder: string;
 };
 
-/**
- * Represents an OBS recorder object.
- */
 class Recorder {
   private _isRecording: boolean = false;
 
@@ -62,8 +40,6 @@ class Recorder {
 
   private _options: RecorderOptionsType;
 
-  private _videoQueue;
-
   private _recorderStartDate = new Date();
 
   private _mainWindow: BrowserWindow;
@@ -72,11 +48,13 @@ class Recorder {
 
   private waitQueue = new WaitQueue<any>();
 
+  private videoProcessQueue;
+
   constructor(mainWindow: BrowserWindow, options: RecorderOptionsType) {
     console.info('[Recorder] Constructing recorder with: ', options);
     this._mainWindow = mainWindow;
+    this.videoProcessQueue = new VideoProcessQueue(mainWindow);
     this._options = options;
-    this.setupVideoProcessingQueue();
     this.createRecordingDirs();
     this.initializeOBS();
     mainWindow.webContents.send('refreshState');
@@ -86,105 +64,18 @@ class Recorder {
     return this._mainWindow;
   }
 
-  private createRecordingDirs() {
-    if (!fs.existsSync(this._options.bufferStorageDir)) {
-      console.log('[Recorder] Creating dir:', this._options.bufferStorageDir);
-      fs.mkdirSync(this._options.bufferStorageDir);
-    } else {
-      console.log('[Recorder] Clean out buffer');
-      this.cleanupBuffer(0);
-    }
-  }
-
-  private async processVideoQueueItem(
-    data: VideoQueueItem,
-    done: Function
-  ): Promise<void> {
-    const videoPath = await this._cutVideo(
-      data.bufferFile,
-      this._options.storageDir,
-      data.filename,
-      data.relativeStart,
-      data.metadata.duration
-    );
-
-    await writeMetadataFile(videoPath, data.metadata);
-
-    // Delete the original buffer video
-    tryUnlinkSync(data.bufferFile);
-
-    done();
-  }
-
-  /**
-   * Setup events on the videoQueue.
-   */
-  private setupVideoProcessingQueue(): void {
-    this._videoQueue = atomicQueue(this.processVideoQueueItem.bind(this), {
-      concurrency: 1,
-    });
-
-    this._videoQueue.on('error', (err: any) => {
-      console.error('[Recorder] Error occured during video processing', err);
-    });
-
-    this._videoQueue.on('idle', () => {
-      console.log('[Recorder] Video processing queue empty, running clean up.');
-
-      // Run the size monitor to ensure we stay within size limit.
-      runSizeMonitor(this._options.storageDir, this._options.maxStorage).then(
-        () => {
-          this.mainWindow.webContents.send('refreshState');
-        }
-      );
-    });
-
-    this._videoQueue.pool.on('start', (data: VideoQueueItem) => {
-      console.log('[Recorder] Processing video', data.bufferFile);
-      this.mainWindow.webContents.send('updateSaveStatus', SaveStatus.Saving);
-    });
-
-    this._videoQueue.pool.on('finish', (_result: any, data: VideoQueueItem) => {
-      console.log('[Recorder] Finished processing video', data.bufferFile);
-      this.mainWindow.webContents.send(
-        'updateSaveStatus',
-        SaveStatus.NotSaving
-      );
-    });
-  }
-
-  /**
-   * Get the value of isRecording.
-   *
-   * @returns {boolean} true if currently recording a game/encounter
-   */
   get isRecording() {
     return this._isRecording;
   }
 
-  /**
-   * Set the value of isRecording.
-   *
-   * @param {boolean} isRecording true if currently recording a game/encounter
-   */
   set isRecording(value) {
     this._isRecording = value;
   }
 
-  /**
-   * Get the value of isRecordingBuffer.
-   *
-   * @returns {boolean} true if currently recording a buffer
-   */
   get isRecordingBuffer() {
     return this._isRecordingBuffer;
   }
 
-  /**
-   * Set the value of isRecordingBuffer.
-   *
-   * @param {boolean} isRecordingBuffer true if currently recording a game/encounter
-   */
   set isRecordingBuffer(value) {
     this._isRecordingBuffer = value;
   }
@@ -235,6 +126,7 @@ class Recorder {
       'updateRecStatus',
       RecStatus.WaitingForWoW
     );
+    
     this.cleanupBuffer(1);
   };
 
@@ -322,8 +214,8 @@ class Recorder {
       this._isRecording = false;
       this._isRecordingBuffer = false;
 
-      const isRaid = activity.category == VideoCategory.Raids;
-      const duration = activity.duration;
+      const isRaid = activity.category === VideoCategory.Raids;
+      const { duration } = activity;
 
       if (duration === null || duration === undefined) {
         console.error('[Recorder] Null or undefined duration');
@@ -343,7 +235,7 @@ class Recorder {
           1000;
 
         if (bufferFile) {
-          this.queueVideo(
+          this.videoProcessQueue.queueVideo(
             bufferFile,
             metadata,
             activity.getFileName(),
@@ -396,32 +288,6 @@ class Recorder {
   };
 
   /**
-   * Queue the video for processing.
-   *
-   * @param {Metadata} metadata the details of the recording
-   */
-  queueVideo = async (
-    bufferFile: string,
-    metadata: Metadata,
-    filename: string,
-    relativeStart: number
-  ) => {
-    // It's a bit hacky that we async wait for 2 seconds for OBS to
-    // finish up with the video file. Maybe this can be done better.
-    setTimeout(async () => {
-      const queueItem: VideoQueueItem = {
-        bufferFile,
-        metadata,
-        filename,
-        relativeStart,
-      };
-
-      console.log('[Recorder] Queuing video for processing', queueItem);
-      this._videoQueue.write(queueItem);
-    }, 2000);
-  };
-
-  /**
    * Clean-up the buffer directory.
    * @params Number of files to leave.
    */
@@ -435,143 +301,14 @@ class Recorder {
     videosToDelete.slice(filesToLeave).forEach((v) => deleteVideo(v.name));
   };
 
-  /**
-   * Shutdown OBS.
-   */
-  shutdown = async () => {
-    if (this._isRecording) {
-      await this.stopOBS();
-      this._isRecording = false;
-    } else if (this._isRecordingBuffer) {
-      this.stopBuffer();
+  private createRecordingDirs() {
+    if (!fs.existsSync(this._options.bufferStorageDir)) {
+      console.info('[Recorder] Creating dir:', this._options.bufferStorageDir);
+      fs.mkdirSync(this._options.bufferStorageDir);
+    } else {
+      console.info('[Recorder] Clean out buffer');
+      this.cleanupBuffer(0);
     }
-
-    obsRecorder.shutdown();
-  };
-
-  /**
-   * Reconfigure the underlying obsRecorder.
-   */
-  reconfigure = async (
-    mainWindow: BrowserWindow,
-    options: RecorderOptionsType
-  ) => {
-    this._mainWindow = mainWindow;
-    this._options = options;
-
-    // User might just have shrunk the size, so run the size monitor.
-    runSizeMonitor(this._options.storageDir, this._options.maxStorage).then(
-      () => {
-        this.mainWindow.webContents.send('refreshState');
-      }
-    );
-
-    if (this._isRecording) {
-      await this.stopOBS();
-      this._isRecording = false;
-    } else if (this._isRecordingBuffer) {
-      this.stopBuffer();
-    }
-
-    obsRecorder.reconfigure(this._options);
-    this.mainWindow.webContents.send('refreshState');
-  };
-
-  /**
-   * Sanitize a filename and replace all invalid characters with a space.
-   *
-   * Multiple consecutive invalid characters will be replaced by a single space.
-   * Multiple consecutive spaces will be replaced by a single space.
-   */
-  private _sanitizeFilename(filename: string): string {
-    return filename
-      .replace(/[<>:"/\|?*]/g, ' ') // Replace all invalid characters with space
-      .replace(/ +/g, ' '); // Replace multiple spaces with a single space
-  }
-
-  /**
-   * Takes an input MP4 file, trims the footage from the start of the video so
-   * that the output is desiredDuration seconds. Some ugly async/await stuff
-   * here. Some interesting implementation details around ffmpeg in comments
-   * below.
-   *
-   * @param {string} initialFile path to initial MP4 file
-   * @param {string} finalDir path to output directory
-   * @param {number} desiredDuration seconds to cut down to
-   * @returns full path of the final video file
-   */
-  private async _cutVideo(
-    initialFile: string,
-    finalDir: string,
-    outputFilename: string | undefined,
-    relativeStart: number,
-    desiredDuration: number
-  ): Promise<string> {
-    const videoFileName = path.basename(initialFile, '.mp4');
-    const videoFilenameSuffix = outputFilename ? ` - ${outputFilename}` : '';
-    const baseVideoFilename = this._sanitizeFilename(
-      videoFileName + videoFilenameSuffix
-    );
-    const finalVideoPath = path.join(finalDir, `${baseVideoFilename}.mp4`);
-
-    return new Promise<string>((resolve) => {
-      // Defensively avoid a negative start time error case.
-      if (relativeStart < 0) {
-        console.log('[Recorder] Video start time was: ', relativeStart);
-        console.log(
-          '[Recorder] Avoiding error by not cutting video from start'
-        );
-        relativeStart = 0;
-      }
-
-      console.log(
-        '[Recorder] Desired duration:',
-        desiredDuration,
-        'Relative start time:',
-        relativeStart
-      );
-
-      // It's crucial that we don't re-encode the video here as that
-      // would spin the CPU and delay the replay being available. I
-      // did try this with re-encoding as it has compression benefits
-      // but took literally ages. My CPU was maxed out for nearly the
-      // same elapsed time as the recording.
-      //
-      // We ensure that we don't re-encode by passing the "-c copy"
-      // option to ffmpeg. Read about it here:
-      // https://superuser.com/questions/377343/cut-part-from-video-file-from-start-position-to-end-position-with-ffmpeg
-      //
-      // This thread has a brilliant summary why we need "-avoid_negative_ts make_zero":
-      // https://superuser.com/questions/1167958/video-cut-with-missing-frames-in-ffmpeg?rq=1
-      ffmpeg(initialFile)
-        .inputOptions([`-ss ${relativeStart}`, `-t ${desiredDuration}`])
-        .outputOptions([
-          `-t ${desiredDuration}`,
-          '-c:v copy',
-          '-c:a copy',
-          '-avoid_negative_ts make_zero',
-        ])
-        .output(finalVideoPath)
-
-        // Handle the end of the FFmpeg cutting.
-        .on('end', async (err: any) => {
-          if (err) {
-            console.log('[Recorder] FFmpeg video cut error (1): ', err);
-            throw new Error('FFmpeg error when cutting video (1)');
-          } else {
-            console.log('[Recorder] FFmpeg cut video succeeded');
-            resolve(finalVideoPath);
-          }
-        })
-
-        // Handle an error with the FFmpeg cutting. Not sure if we
-        // need this as well as the above but being careful.
-        .on('error', (err: any) => {
-          console.log('[Recorder] FFmpeg video cut error (2): ', err);
-          throw new Error('FFmpeg error when cutting video (2)');
-        })
-        .run();
-    });
   }
 
   private initializeOBS() {
